re-enable elu and celu tests
