+++ src/mlpack/tests/ann/activation_functions_test.cpp
@@ -36,7 +36,7 @@
 #include <mlpack/methods/ann/activation_functions/tanh_exponential_function.hpp>
 #include <mlpack/methods/ann/activation_functions/silu_function.hpp>
 
-#include "catch.hpp"
+#include "../catch.hpp"
 
 using namespace mlpack;
 using namespace mlpack::ann;
@@ -130,53 +130,6 @@ void CheckInverseCorrect(const arma::colvec input)
 }
 
 /**
- * Implementation of the HardTanH activation function test. The function is
- * implemented as a HardTanH Layer in hard_tanh.hpp
- *
- * @param input Input data used for evaluating the HardTanH activation function.
- * @param target Target data used to evaluate the HardTanH activation.
- *
-void CheckHardTanHActivationCorrect(const arma::colvec input,
-                                    const arma::colvec target)
-{
-  HardTanH<> htf;
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  htf.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the HardTanH activation function derivative test. The
- * derivative is implemented as HardTanH Layer in hard_tanh.hpp
- *
- * @param input Input data used for evaluating the HardTanH activation
- * function.
- * @param target Target data used to evaluate the HardTanH activation.
- *
-void CheckHardTanHDerivativeCorrect(const arma::colvec input,
-                                    const arma::colvec target)
-{
-  HardTanH<> htf;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  htf.Backward(input, error, derivatives);
-
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
  * Implementation of the LeakyReLU activation function test. The function is
  * implemented as LeakyReLU layer in the file leaky_relu.hpp
  *
@@ -230,21 +183,24 @@ void CheckLeakyReLUDerivativeCorrect(const arma::colvec input,
  *
  * @param input Input data used for evaluating the ELU activation function.
  * @param target Target data used to evaluate the ELU activation.
- *
+ */
 void CheckELUActivationCorrect(const arma::colvec input,
-                               const arma::colvec target)
+    const arma::colvec target)
 {
-  // Initialize ELU object with alpha = 1.0.
-  ELU<> lrf(1.0);
+  // Initialize ELU object with alpha = 1.
+  ELU module(1);
+  module.Training() = true;
+  module.InputDimensions() = std::vector<size_t>({ input.n_elem });
+  module.ComputeOutputDimensions();
 
   // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  lrf.Forward(input, activations);
+  arma::colvec activations(input.n_elem);
+  module.Forward(input, activations);
   for (size_t i = 0; i < activations.n_elem; ++i)
   {
     REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
   }
-}*/
+}
 
 /**
  * Implementation of the ELU activation function derivative test. The function
@@ -252,255 +208,139 @@ void CheckELUActivationCorrect(const arma::colvec input,
  *
  * @param input Input data used for evaluating the ELU activation function.
  * @param target Target data used to evaluate the ELU activation.
- *
+ */
 void CheckELUDerivativeCorrect(const arma::colvec input,
-                               const arma::colvec target)
+    const arma::colvec target)
 {
   // Initialize ELU object with alpha = 1.0.
-  ELU<> lrf(1.0);
+  ELU module(1);
+  module.Training() = true;
+  module.InputDimensions() = std::vector<size_t>({ input.n_elem });
+  module.ComputeOutputDimensions();
 
   // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives, activations;
+  arma::colvec derivatives(input.n_elem), activations(input.n_elem);
 
   // This error vector will be set to 1 to get the derivatives.
   arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  lrf.Forward(input, activations);
-  lrf.Backward(activations, error, derivatives);
+  module.Forward(input, activations);
+  module.Backward(activations, error, derivatives);
   for (size_t i = 0; i < derivatives.n_elem; ++i)
   {
     REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
   }
-}*/
+}
 
 /**
- * Implementation of the PReLU activation function test. The function
- * is implemented as PReLU layer in the file parametric_relu.hpp.
+ * Implementation of the CELU activation function test. The function is
+ * implemented as CELU layer in the file celu.hpp.
  *
- * @param input Input data used for evaluating the PReLU activation
- *   function.
- * @param target Target data used to evaluate the PReLU activation.
- *
-void CheckPReLUActivationCorrect(const arma::colvec input,
-                                 const arma::colvec target)
+ * @param input Input data used for evaluating the CELU activation function.
+ * @param target Target data used to evaluate the CELU activation.
+ */
+void CheckCELUActivationCorrect(const arma::colvec input,
+    const arma::colvec target)
 {
-  PReLU<> prelu;
+  // Initialize CELU object with alpha = 1.
+  CELU module(1);
+  module.Training() = true;
+  module.InputDimensions() = std::vector<size_t>({ input.n_elem });
+  module.ComputeOutputDimensions();
 
   // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  prelu.Forward(input, activations);
+  arma::colvec activations(input.n_elem);
+  module.Forward(input, activations);
   for (size_t i = 0; i < activations.n_elem; ++i)
   {
     REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
   }
-}*/
+}
 
 /**
- * Implementation of the PReLU activation function derivative test.
- * The function is implemented as PReLU layer in the file
- * parametric_relu.hpp
+ * Implementation of the CELU activation function derivative test. The function
+ * is implemented as CELU layer in the file celu.hpp.
  *
- * @param input Input data used for evaluating the PReLU activation
- *   function.
- * @param target Target data used to evaluate the PReLU activation.
- *
-void CheckPReLUDerivativeCorrect(const arma::colvec input,
-                                 const arma::colvec target)
+ * @param input Input data used for evaluating the CELU activation function.
+ * @param target Target data used to evaluate the CELU activation.
+ */
+void CheckCELUDerivativeCorrect(const arma::colvec input,
+    const arma::colvec target)
 {
-  PReLU<> prelu;
+  // Initialize CELU object with alpha = 1.0.
+  CELU module(1);
+  module.Training() = true;
+  module.InputDimensions() = std::vector<size_t>({ input.n_elem });
+  module.ComputeOutputDimensions();
 
   // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives;
+  arma::colvec derivatives(input.n_elem), activations(input.n_elem);
 
   // This error vector will be set to 1 to get the derivatives.
   arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  prelu.Backward(input, error, derivatives);
+  module.Forward(input, activations);
+  module.Backward(activations, error, derivatives);
   for (size_t i = 0; i < derivatives.n_elem; ++i)
   {
     REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
   }
-}*/
-
-/**
- * Implementation of the PReLU activation function gradient test.
- * The function is implemented as PReLU layer in the file
- * parametric_relu.hpp
- *
- * @param input Input data used for evaluating the PReLU activation
- *   function.
- * @param target Target data used to evaluate the PReLU gradient.
- *
-void CheckPReLUGradientCorrect(const arma::colvec input,
-                               const arma::colvec target)
-{
-  PReLU<> prelu;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec gradient;
-
-  // This error vector will be set to 1 to get the gradient.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  prelu.Gradient(input, error, gradient);
-  REQUIRE(gradient.n_rows == 1);
-  REQUIRE(gradient.n_cols == 1);
-  REQUIRE(gradient(0) == Approx(target(0)).epsilon(1e-5));
-}*/
-
-/**
- * Implementation of the Hard Shrink activation function test. The function is
- * implemented as Hard Shrink layer in the file hardshrink.hpp
- *
- * @param input Input data used for evaluating the Hard Shrink activation function.
- * @param target Target data used to evaluate the Hard Shrink activation.
- *
-void CheckHardShrinkActivationCorrect(const arma::colvec input,
-                                      const arma::colvec target)
-{
-  HardShrink<> hardshrink;
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  hardshrink.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the HardShrink activation function derivative test.
- * The derivative function is implemented as HardShrink layer in the file
- * hardshrink.hpp
- *
- * @param input Input data used for evaluating the HardShrink activation
- * function.
- * @param target Target data used to evaluate the HardShrink activation.
- *
-void CheckHardShrinkDerivativeCorrect(const arma::colvec input,
-                                      const arma::colvec target)
-{
-  HardShrink<> hardshrink;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  hardshrink.Backward(input, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Soft Shrink activation function test. The function is
- * implemented as Soft Shrink layer in the file softshrink.hpp.
- *
- * @param input Input data used for evaluating the Soft Shrink activation
- * function.
- * @param target Target data used to evaluate the Soft Shrink activation.
- *
-void CheckSoftShrinkActivationCorrect(const arma::colvec input,
-                                      const arma::colvec target)
-{
-  SoftShrink<> softshrink;
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  softshrink.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Soft Shrink activation function derivative test.
- * The derivative function is implemented as Soft Shrink layer in the file
- * softshrink.hpp
- *
- * @param input Input data used for evaluating the Soft Shrink activation
- * function.
- * @param target Target data used to evaluate the Soft Shrink activation.
- *
-void CheckSoftShrinkDerivativeCorrect(const arma::colvec input,
-                                      const arma::colvec target)
-{
-  SoftShrink<> softshrink;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  softshrink.Backward(input, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
+}
 
 /**
  * Simple SELU activation test to check whether the mean and variance remain
  * invariant after passing normalized inputs through the function.
- *
+ */
 TEST_CASE("SELUFunctionNormalizedTest", "[ActivationFunctionsTest]")
 {
   arma::mat input = arma::randn<arma::mat>(1000, 1);
-
-  arma::mat output;
-
+  arma::mat output(arma::size(input));
+  
   SELU selu;
-
   selu.Forward(input, output);
 
   REQUIRE(arma::as_scalar(arma::abs(arma::mean(input) -
       arma::mean(output))) <= 0.1);
-
   REQUIRE(arma::as_scalar(arma::abs(arma::var(input) -
       arma::var(output))) <= 0.1);
-}*/
+}
 
 /**
  * Simple SELU activation test to check whether the mean and variance
  * vary significantly after passing unnormalized inputs through the function.
- *
+ */
 TEST_CASE("SELUFunctionUnnormalizedTest", "[ActivationFunctionsTest]")
 {
   const arma::colvec input("5.96402758 0.9966824 0.99975321 1 \
                             7.76159416 -0.76159416 0.96402758 8");
-
-  arma::mat output;
-
+  arma::mat output(arma::size(input));
+  
   SELU selu;
-
   selu.Forward(input, output);
 
   REQUIRE(arma::as_scalar(arma::abs(arma::mean(input) -
       arma::mean(output))) >= 0.1);
-
   REQUIRE(arma::as_scalar(arma::abs(arma::var(input) -
       arma::var(output))) >= 0.1);
-}*/
+}
 
 /**
  * Simple SELU derivative test to check whether the derivatives
  * produced by the activation function are correct.
  *
- *
+ */
 TEST_CASE("SELUFunctionDerivativeTest", "[ActivationFunctionsTest]")
 {
   arma::mat input = arma::ones<arma::mat>(1000, 1);
-
   arma::mat error = arma::ones<arma::mat>(input.n_elem, 1);
-
-  arma::mat derivatives, activations;
+  arma::mat derivatives(arma::size(input)), activations(arma::size(input));
 
   SELU selu;
+  selu.Training() = true;
+  selu.InputDimensions() = std::vector<size_t>({ input.n_elem });
+  selu.ComputeOutputDimensions();
 
   selu.Forward(input, activations);
   selu.Backward(activations, error, derivatives);
-
+  
   REQUIRE(arma::as_scalar(arma::abs(arma::mean(derivatives) -
       selu.Lambda())) <= 10e-4);
 
@@ -511,245 +351,7 @@ TEST_CASE("SELUFunctionDerivativeTest", "[ActivationFunctionsTest]")
 
   REQUIRE(arma::as_scalar(arma::abs(arma::mean(derivatives) -
       selu.Lambda() * selu.Alpha() - arma::mean(activations))) <= 10e-4);
-}*/
-
-/**
- * Implementation of the CELU activation function test. The function is
- * implemented as CELU layer in the file celu.hpp.
- *
- * @param input Input data used for evaluating the CELU activation function.
- * @param target Target data used to evaluate the CELU activation.
- *
-void CheckCELUActivationCorrect(const arma::colvec input,
-                                const arma::colvec target)
-{
-  // Initialize CELU object with alpha = 1.0.
-  CELU<> lrf(1.0);
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  lrf.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the CELU activation function derivative test. The function
- * is implemented as CELU layer in the file celu.hpp.
- *
- * @param input Input data used for evaluating the CELU activation function.
- * @param target Target data used to evaluate the CELU activation.
- *
-void CheckCELUDerivativeCorrect(const arma::colvec input,
-                                const arma::colvec target)
-{
-  // Initialize CELU object with alpha = 1.0.
-  CELU<> lrf(1.0);
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives, activations;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  lrf.Forward(input, activations);
-  lrf.Backward(activations, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the ISRLU activation function test. The function is
- * implemented as ISRLU layer in the file isrlu.hpp.
- *
- * @param input Input data used for evaluating the ISRLU activation function.
- * @param target Target data used to evaluate the ISRLU activation.
- *
-void CheckISRLUActivationCorrect(const arma::colvec input,
-                                 const arma::colvec target)
-{
-  // Initialize ISRLU object with alpha = 1.0.
-  ISRLU<> lrf(1.0);
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  lrf.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the ISRLU activation function derivative test. The function
- * is implemented as ISRLU layer in the file isrlu.hpp.
- *
- * @param input Input data used for evaluating the ISRLU activation function.
- * @param target Target data used to evaluate the ISRLU activation.
- *
-void CheckISRLUDerivativeCorrect(const arma::colvec input,
-                                 const arma::colvec target)
-{
-  // Initialize ISRLU object with alpha = 1.0.
-  ISRLU<> lrf(1.0);
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives, activations;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  lrf.Forward(input, activations);
-  lrf.Backward(activations, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Softmin activation function test. The function is
- * implemented as Softmin layer in the file softmin.hpp.
- *
- * @param input Input data used for evaluating the Softmin activation function.
- * @param target Target data used to evaluate the Softmin activation.
- *
-void CheckSoftminActivationCorrect(const arma::colvec input,
-                                   const arma::colvec target)
-{
-  // Initialize Softmin object.
-  Softmin<> softmin;
-
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  softmin.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Softmin activation function derivative test.
- * The function is implemented as Softmin layer in the file softmin.hpp.
- *
- * @param input Input data used for evaluating the Softmin activation function.
- * @param target Target data used to evaluate the Softmin activation.
- *
-void CheckSoftminDerivativeCorrect(const arma::colvec input,
-                                   const arma::colvec target)
-{
-  // Initialize Softmin object.
-  Softmin<> softmin;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives, activations;
-
-  // This error vector will be set to [[1.0],[0.0],[1.0],[0.0]]
-  // to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  error(1) = 0.0;
-  error(3) = 0.0;
-  softmin.Forward(input, activations);
-  softmin.Backward(activations, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Flatten T Swish activation function test. The function is
- * implemented as Flatten T Swish layer in the file flatten_t_swish.hpp.
- *
- * @param input Input data used for evaluating the Flatten T Swish activation
- *     function.
- * @param target Target data used to evaluate the Flatten T Swish activation.
- *
-void CheckFlattenTSwishActivationCorrect(const arma::colvec input,
-                                         const arma::colvec target)
-{
-  FlattenTSwish<> fts(0.4);
-  arma::colvec activations;
-
-  fts.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the Softmin activation function derivative test.
- * The function is implemented as Softmin layer in the file softmin.hpp.
- *
- * @param input Input data used for evaluating the Softmin activation function.
- * @param target Target data used to evaluate the Softmin activation.
- *
-void CheckFlattenTSwishDerivateCorrect(const arma::colvec input,
-                                       const arma::colvec target)
-{
-  FlattenTSwish<> fts;
-
-  // Set the error to 1 to get the actual derivative.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-
-  arma::colvec derivate;
-  fts.Backward(input, error, derivate);
-  for (size_t i = 0; i < derivate.n_elem; ++i)
-  {
-    REQUIRE(derivate.at(i) == Approx(target.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Implementation of the ReLU6 activation function derivative test. The function
- * is implemented as ReLU6 layer in the file relu6.hpp.
- *
- * @param input Input data used for evaluating the ReLU6 activation function.
- * @param target Target data used to evaluate the ReLU6 activation.
- *
-void CheckReLU6Correct(const arma::colvec input,
-                       const arma::colvec ActivationTarget,
-                       const arma::colvec DerivativeTarget)
-{
-  // Initialize ReLU6 object.
-  ReLU6<> relu6;
-
-  // Test the calculation of the derivatives using the entire vector as input.
-  arma::colvec derivatives, activations;
-
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(input.n_elem);
-  relu6.Forward(input, activations);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) == Approx(ActivationTarget.at(i)).epsilon(1e-5));
-  }
-  relu6.Backward(activations, error, derivatives);
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) == Approx(DerivativeTarget.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Basic test of the ReLU6 function.
- *
-TEST_CASE("ReLU6FunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec activationData("-2.0 3.0 0.0 6.0 24.0");
-
-  // desiredActivations taken from PyTorch.
-  const arma::colvec desiredActivations("0.0 3.0 0.0 6.0 6.0");
-
-  // desiredDerivatives taken from PyTorch.
-  const arma::colvec desiredDerivatives("0.0 1.0 0.0 0.0 0.0");
-
-  CheckReLU6Correct(activationData, desiredActivations, desiredDerivatives);
-}*/
+}
 
 /**
  * Basic test of the tanh function.
@@ -846,23 +448,8 @@ TEST_CASE("LeakyReLUFunctionTest", "[ActivationFunctionsTest]")
 }
 
 /**
- * Basic test of the HardTanH function.
- *
-TEST_CASE("HardTanHFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-1 1 1 -1 \
-                                         1 -1 1 0");
-
-  const arma::colvec desiredDerivatives("0 0 0 0 \
-                                         1 1 0 1");
-
-  CheckHardTanHActivationCorrect(activationData, desiredActivations);
-  CheckHardTanHDerivativeCorrect(activationData, desiredDerivatives);
-}*/
-
-/**
  * Basic test of the ELU function.
- *
+ */
 TEST_CASE("ELUFunctionTest", "[ActivationFunctionsTest]")
 {
   const arma::colvec desiredActivations("-0.86466471 3.2 4.5 -1.0 \
@@ -873,7 +460,23 @@ TEST_CASE("ELUFunctionTest", "[ActivationFunctionsTest]")
 
   CheckELUActivationCorrect(activationData, desiredActivations);
   CheckELUDerivativeCorrect(activationData, desiredDerivatives);
-}*/
+}
+
+/**
+ * Basic test of the CELU activation function.
+ */
+TEST_CASE("CELUFunctionTest", "[ActivationFunctionsTest]")
+{
+  const arma::colvec desiredActivations("-0.86466472 3.2 4.5 \
+                                         -1 1 -0.63212056 2 0");
+
+  const arma::colvec desiredDerivatives("0.42119275 1 1 \
+                                         0.36787944 1 \
+                                         0.5314636 1 1");
+
+  CheckCELUActivationCorrect(activationData, desiredActivations);
+  CheckCELUDerivativeCorrect(desiredActivations, desiredDerivatives);
+}
 
 /**
  * Basic test of the softplus function.
@@ -895,73 +498,6 @@ TEST_CASE("SoftplusFunctionTest", "[ActivationFunctionsTest]")
                                            desiredDerivatives);
   CheckInverseCorrect<SoftplusFunction>(desiredActivations);
 }
-
-/**
- * Basic test of the PReLU function.
- *
-TEST_CASE("PReLUFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-0.06 3.2 4.5 -3.006 \
-                                         1 -0.03 2 0");
-
-  const arma::colvec desiredDerivatives("0.03 1 1 0.03 \
-                                         1 0.03 1 1");
-  const arma::colvec desiredGradient("-103.2");
-
-  CheckPReLUActivationCorrect(activationData, desiredActivations);
-  CheckPReLUDerivativeCorrect(desiredActivations, desiredDerivatives);
-  CheckPReLUGradientCorrect(activationData, desiredGradient);
-}*/
-
-/**
- * Basic test of the CReLU function.
- *
-TEST_CASE("CReLUFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("0 3.2 4.5 0 \
-                                         1 0 2 0 2 0 0 \
-                                         100.2 0 1 0 0");
-
-  const arma::colvec desiredDerivatives("0 0 0 0 \
-                                         0 0 0 0");
-  CReLU<> crelu;
-  // Test the activation function using the entire vector as input.
-  arma::colvec activations;
-  crelu.Forward(activationData, activations);
-  arma::colvec derivatives;
-  // This error vector will be set to 1 to get the derivatives.
-  arma::colvec error = arma::ones<arma::colvec>(desiredActivations.n_elem);
-  crelu.Backward(desiredActivations, error, derivatives);
-  for (size_t i = 0; i < activations.n_elem; ++i)
-  {
-    REQUIRE(activations.at(i) ==
-        Approx(desiredActivations.at(i)).epsilon(1e-5));
-  }
-  for (size_t i = 0; i < derivatives.n_elem; ++i)
-  {
-    REQUIRE(derivatives.at(i) ==
-        Approx(desiredDerivatives.at(i)).epsilon(1e-5));
-  }
-}*/
-
-/**
- * Basic test of the swish function.
- *
-TEST_CASE("SwishFunctionTest", "[ActivationFunctionsTest]")
-{
-  // Hand-calculated values using Python interpreter.
-  const arma::colvec desiredActivations("-0.238405 3.07466 4.45055 \
-                                         -3.05183208657e-42 0.731058 -0.26894 \
-                                         1.76159 0");
-
-  const arma::colvec desiredDerivatives("0.3819171 1.0856295 1.039218 \
-                                         0.5 0.83540367 0.3671335 1.073787\
-                                         0.5");
-
-  CheckActivationCorrect<SwishFunction>(activationData, desiredActivations);
-  CheckDerivativeCorrect<SwishFunction>(desiredActivations,
-                                        desiredDerivatives);
-}*/
 
 /**
  * Basic test of the hard sigmoid function.
@@ -1048,21 +584,6 @@ TEST_CASE("GELUFunctionTest", "[ActivationFunctionsTest]")
 }
 
 /**
- * Basic test of the Hard Shrink function.
- *
-TEST_CASE("HardShrinkFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-2 3.2 4.5 -100.2 1 -1 2 0");
-
-  const arma::colvec desiredDerivatives("1 1 1 1 1 1 1 0");
-
-  CheckHardShrinkActivationCorrect(activationData,
-                                   desiredActivations);
-  CheckHardShrinkDerivativeCorrect(desiredActivations,
-                                   desiredDerivatives);
-}*/
-
-/**
  * Basic test of the Elliot function.
  */
 TEST_CASE("ElliotFunctionTest", "[ActivationFunctionsTest]")
@@ -1101,53 +622,6 @@ TEST_CASE("ElishFunctionTest", "[ActivationFunctionsTest]")
   CheckDerivativeCorrect<ElishFunction>(desiredActivations,
                                         desiredDerivatives);
 }
-
-/**
- * Basic test of the Soft Shrink function.
- *
-TEST_CASE("SoftShrinkFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-1.5 2.7 4 -99.7 0.5 -0.5 1.5 0");
-
-  const arma::colvec desiredDerivatives("1 1 1 1 1 1 1 0");
-
-  CheckSoftShrinkActivationCorrect(activationData,
-                                   desiredActivations);
-  CheckSoftShrinkDerivativeCorrect(desiredActivations,
-                                   desiredDerivatives);
-}*/
-
-/**
- * Basic test of the CELU activation function.
- *
-TEST_CASE("CELUFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-0.86466472 3.2 4.5 \
-                                         -1 1 -0.63212056 2 0");
-
-  const arma::colvec desiredDerivatives("0.42119275 1 1 \
-                                         0.36787944 1 \
-                                         0.5314636 1 1");
-
-  CheckCELUActivationCorrect(activationData, desiredActivations);
-  CheckCELUDerivativeCorrect(desiredActivations, desiredDerivatives);
-}*/
-
-/**
- * Basic test of the ISRLU activation function.
- *
-TEST_CASE("ISRLUFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec desiredActivations("-0.89442719 3.2 4.5 \
-                                         -0.99995020 1 -0.70710678 2 0");
-
-  const arma::colvec desiredDerivatives("0.41408666 1 1 \
-                                         0.35357980 1 \
-                                         0.54433105 1 1");
-
-  CheckISRLUActivationCorrect(activationData, desiredActivations);
-  CheckISRLUDerivativeCorrect(activationData, desiredDerivatives);
-}*/
 
 /**
  * Basic test of the inverse quadratic function.
@@ -1274,26 +748,6 @@ TEST_CASE("GaussianFunctionTest", "[ActivationFunctionsTest]")
 }
 
 /**
- * Basic test of the Softmin function.
- *
-TEST_CASE("SoftminFunctionTest", "[ActivationFunctionsTest]")
-{
-  const arma::colvec activationData("4.2 2.4 7.0 6.4");
-
-  // Hand-calculated Values.
-  const arma::colvec desiredActivations("0.1384799751 0.8377550303 \
-                                         0.008420976 0.0153440186");
-
-  const arma::colvec desiredDerivatives("0.1181371351 -0.12306701070 \
-                                         0.0071839266 -0.0022540509");
-
-  CheckSoftminActivationCorrect(activationData,
-                                desiredActivations);
-  CheckSoftminDerivativeCorrect(activationData,
-                                desiredDerivatives);
-}*/
-
-/**
  * Basic test of the Hard Swish function.
  */
 TEST_CASE("HardSwishFunctionTest", "[ActivationFunctionsTest]")
@@ -1357,25 +811,3 @@ TEST_CASE("SILUFunctionTest", "[ActivationFunctionsTest]")
   CheckActivationCorrect<SILUFunction>(activationData, desiredActivation);
   CheckDerivativeCorrect<SILUFunction>(desiredActivation, desiredDerivate);
 }
-
-/**
- * Basic test of Flatten T Swish function.
- *
-TEST_CASE("FlattenTSwishFunctionTest", "[ActivationFunctionsTest]")
-{
-  // Random Value.
-  arma::colvec input("-4.0 -1.0 2 3 4 5 6");
-
-  // Hand Calculated and using PyTorch.
-  arma::colvec desiredActivation(
-      "0.4000000059604645 0.4000000059604645 2.1615941524505615 \
-       3.2577223777770996 4.328054904937744 5.3665361404418945 \
-       6.385164737701416");
-
-  // Hand Calculated and using PyTorch.
-  arma::colvec desiredDerivation("0.694792 0.694792 1.096893 1.079178 1.042602 \
-                                  1.020182 1.009048");
-
-  CheckFlattenTSwishActivationCorrect(input, desiredActivation);
-  CheckFlattenTSwishDerivateCorrect(desiredActivation, desiredDerivation);
-}*/
