+++ src/mlpack/tests/ann/ann_dist_test.cpp
@@ -16,8 +16,8 @@
 #include <mlpack/methods/ann/dists/normal_distribution.hpp>
 #include <mlpack/methods/ann/init_rules/random_init.hpp>
 
-#include "catch.hpp"
-#include "test_catch_tools.hpp"
+#include "../catch.hpp"
+#include "../test_catch_tools.hpp"
 
 #include <mlpack/methods/ann/activation_functions/logistic_function.hpp>
