+++ src/mlpack/tests/ann/convolution_test.cpp
@@ -17,9 +17,9 @@
 #include <mlpack/methods/ann/convolution_rules/fft_convolution.hpp>
 #include <mlpack/methods/ann/convolution_rules/svd_convolution.hpp>
 
-#include "serialization.hpp"
-#include "catch.hpp"
-#include "test_catch_tools.hpp"
+#include "../serialization.hpp"
+#include "../catch.hpp"
+#include "../test_catch_tools.hpp"
 
 using namespace mlpack;
 using namespace mlpack::ann;
