+++ src/mlpack/tests/ann/loss_functions_test.cpp
@@ -40,8 +40,8 @@
 #include <mlpack/methods/ann/init_rules/nguyen_widrow_init.hpp>
 #include <mlpack/methods/ann/ffn.hpp>
 
-#include "catch.hpp"
-#include "test_catch_tools.hpp"
+#include "../catch.hpp"
+#include "../test_catch_tools.hpp"
 #include "ann_test_tools.hpp"
 
 using namespace mlpack;
