+++ src/mlpack/tests/ann/not_adapted/ann_layer_test.cpp
@@ -22,10 +22,10 @@
 #include <mlpack/methods/ann/loss_functions/binary_cross_entropy_loss.hpp>
 #include <mlpack/methods/ann/ffn.hpp>
 
-#include "test_catch_tools.hpp"
-#include "catch.hpp"
+#include "../test_catch_tools.hpp"
+#include "../catch.hpp"
+#include "../serialization.hpp"
 #include "ann_test_tools.hpp"
-#include "serialization.hpp"
 
 using namespace mlpack;
 using namespace mlpack::ann;
@@ -238,178 +238,6 @@ TEST_CASE("ConstantLayerParametersTest", "[ANNLayerTest]")
   REQUIRE(layer.OutSize() == 7);
 }*/
 
-/**
- * Simple dropout module test.
- */
-TEST_CASE("SimpleDropoutLayerTest", "[ANNLayerTest]")
-{
-  // Initialize the probability of setting a value to zero.
-  const double p = 0.2;
-
-  // Initialize the input parameter.
-  arma::mat input(1000, 1);
-  input.fill(1 - p);
-
-  Dropout module(p);
-  module.Training() = true;
-
-  // Test the Forward function.
-  arma::mat output;
-  module.Forward(input, output);
-  REQUIRE(arma::as_scalar(arma::abs(arma::mean(output) - (1 - p))) <= 0.05);
-
-  // Test the Backward function.
-  arma::mat delta;
-  module.Backward(input, input, delta);
-  REQUIRE(arma::as_scalar(arma::abs(arma::mean(delta) - (1 - p))) <= 0.05);
-
-  // Test the Forward function.
-  module.Training() = false;
-  module.Forward(input, output);
-  REQUIRE(arma::accu(input) == arma::accu(output));
-}
-
-/**
- * Perform dropout x times using ones as input, sum the number of ones and
- * validate that the layer is producing approximately the correct number of
- * ones.
- */
-TEST_CASE("DropoutProbabilityTest", "[ANNLayerTest]")
-{
-  arma::mat input = arma::ones(1500, 1);
-  const size_t iterations = 10;
-
-  double probability[5] = { 0.1, 0.3, 0.4, 0.7, 0.8 };
-  for (size_t trial = 0; trial < 5; ++trial)
-  {
-    double nonzeroCount = 0;
-    for (size_t i = 0; i < iterations; ++i)
-    {
-      Dropout module(probability[trial]);
-      module.Training() = true;
-
-      arma::mat output;
-      module.Forward(input, output);
-
-      // Return a column vector containing the indices of elements of X that
-      // are non-zero, we just need the number of non-zero values.
-      arma::uvec nonzero = arma::find(output);
-      nonzeroCount += nonzero.n_elem;
-    }
-    const double expected = input.n_elem * (1 - probability[trial]) *
-        iterations;
-    const double error = fabs(nonzeroCount - expected) / expected;
-
-    REQUIRE(error <= 0.15);
-  }
-}
-
-/*
- * Perform dropout with probability 1 - p where p = 0, means no dropout.
- */
-TEST_CASE("NoDropoutTest", "[ANNLayerTest]")
-{
-  arma::mat input = arma::ones(1500, 1);
-  Dropout module(0);
-  module.Training() = true;
-
-  arma::mat output;
-  module.Forward(input, output);
-
-  REQUIRE(arma::accu(output) == arma::accu(input));
-}
-
-/*
- * Perform test to check whether mean and variance remain nearly same
- * after AlphaDropout.
- */
-TEST_CASE("SimpleAlphaDropoutLayerTest", "[ANNLayerTest]")
-{
-  // Initialize the probability of setting a value to alphaDash.
-  const double p = 0.2;
-
-  // Initialize the input parameter having a mean nearabout 0
-  // and variance nearabout 1.
-  arma::mat input = arma::randn<arma::mat>(1000, 1);
-
-  AlphaDropout module(p);
-  module.Training() = true;
-
-  // Test the Forward function when training phase.
-  arma::mat output(arma::size(input));
-  module.Forward(input, output);
-  // Check whether mean remains nearly same.
-  REQUIRE(arma::as_scalar(arma::abs(arma::mean(input) - arma::mean(output))) <=
-      0.1);
-
-  // Check whether variance remains nearly same.
-  REQUIRE(arma::as_scalar(arma::abs(arma::var(input) - arma::var(output))) <=
-      0.1);
-
-  // Test the Backward function when training phase.
-  arma::mat delta;
-  module.Backward(input, input, delta);
-  REQUIRE(arma::as_scalar(arma::abs(arma::mean(delta) - 0)) <= 0.05);
-
-  // Test the Forward function when testing phase.
-  module.Training() = false;
-  module.Forward(input, output);
-  REQUIRE(arma::accu(input) == arma::accu(output));
-}
-
-/**
- * Perform AlphaDropout x times using ones as input, sum the number of ones
- * and validate that the layer is producing approximately the correct number
- * of ones.
- */
-TEST_CASE("AlphaDropoutProbabilityTest", "[ANNLayerTest]")
-{
-  arma::mat input = arma::ones(1500, 1);
-  const size_t iterations = 10;
-
-  double probability[5] = { 0.1, 0.3, 0.4, 0.7, 0.8 };
-  for (size_t trial = 0; trial < 5; ++trial)
-  {
-    double nonzeroCount = 0;
-    for (size_t i = 0; i < iterations; ++i)
-    {
-      AlphaDropout module(probability[trial]);
-      module.Training() = true;
-
-      arma::mat output(arma::size(input));
-      module.Forward(input, output);
-
-      // Return a column vector containing the indices of elements of X
-      // that are not alphaDash, we just need the number of
-      // nonAlphaDash values.
-      arma::uvec nonAlphaDash = arma::find(module.Mask());
-      nonzeroCount += nonAlphaDash.n_elem;
-    }
-
-    const double expected = input.n_elem * (1-probability[trial]) * iterations;
-
-    const double error = fabs(nonzeroCount - expected) / expected;
-
-    REQUIRE(error <= 0.15);
-  }
-}
-
-/**
- * Perform AlphaDropout with probability 1 - p where p = 0,
- * means no AlphaDropout.
- */
-TEST_CASE("NoAlphaDropoutTest", "[ANNLayerTest]")
-{
-  arma::mat input = arma::ones(1500, 1);
-  AlphaDropout module(0);
-  module.Training() = false;
-
-  arma::mat output;
-  module.Forward(input, output);
-
-  REQUIRE(arma::accu(output) == arma::accu(input));
-}
-
 // /**
 //  * Simple linear module test.
 //  */
@@ -494,121 +322,6 @@ TEST_CASE("NoAlphaDropoutTest", "[ANNLayerTest]")
 //   REQUIRE(CheckGradient(function) <= 1e-4);
 // }
 
-/**
- * Simple Linear3D layer test.
- */
-TEST_CASE("SimpleLinear3DLayerTest", "[ANNLayerTest]")
-{
-  const size_t inSize = 4;
-  const size_t outSize = 1;
-  const size_t nPoints = 2;
-  const size_t batchSize = 1;
-  arma::mat input, output, delta;
-
-  // Create a Linear3D layer outside of a network, and then set its memory.
-  Linear3D module(outSize);
-  module.InputDimensions() = std::vector<size_t>({ 4, 2 });
-  module.ComputeOutputDimensions();
-  arma::mat weights(module.WeightSize(), 1);
-  module.SetWeights(weights.memptr());
-
-  module.Parameters().randu();
-
-  // Test the Forward function.
-  input = arma::zeros(inSize * nPoints, batchSize);
-  output.set_size(outSize * nPoints, batchSize);
-  module.Forward(input, output);
-  REQUIRE(arma::accu(module.Bias())
-      == Approx(arma::accu(output) / (nPoints * batchSize)).epsilon(1e-3));
-
-  // Test the Backward function.
-  delta.set_size(input.n_rows, input.n_cols);
-  output.zeros();
-  module.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 0);
-}
-
-/**
- * Jacobian Linear3D module test.
- */
-TEST_CASE("JacobianLinear3DLayerTest", "[ANNLayerTest]")
-{
-  for (size_t i = 0; i < 5; ++i)
-  {
-    const size_t inSize = math::RandInt(2, 10);
-    const size_t outSize = math::RandInt(2, 10);
-    const size_t nPoints = math::RandInt(2, 10);
-    const size_t batchSize = 1;
-
-    arma::mat input;
-    input.set_size(inSize * nPoints, batchSize);
-
-    // Create a Linear3D layer outside a network and initialize its memory.
-    Linear3D module(outSize);
-    module.InputDimensions() = std::vector<size_t>({ inSize, nPoints });
-    module.ComputeOutputDimensions();
-    arma::mat weights(module.WeightSize(), 1);
-    module.SetWeights(weights.memptr());
-
-    module.Parameters().randu();
-
-    double error = JacobianTest(module, input);
-    REQUIRE(error <= 1e-5);
-  }
-}
-
-/**
- * Simple Gradient test for Linear3D layer.
- */
-TEST_CASE("GradientLinear3DLayerTest", "[ANNLayerTest]")
-{
-  // Linear function gradient instantiation.
-  struct GradientFunction
-  {
-    GradientFunction() :
-        inSize(4),
-        outSize(1),
-        nPoints(2),
-        batchSize(4)
-    {
-      input = arma::randu(inSize * nPoints, batchSize);
-      target = arma::zeros(outSize * nPoints, batchSize);
-      target(0, 0) = 1;
-      target(0, 3) = 1;
-      target(1, 1) = 1;
-      target(1, 2) = 1;
-
-      model = new FFN<MeanSquaredError, RandomInitialization>();
-      model->ResetData(input, target);
-      model->Add<Linear3D>(outSize);
-      model->InputDimensions() = std::vector<size_t>{ 4, 2 };
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 1);
-      model->Gradient(model->Parameters(), 0, gradient, 1);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<MeanSquaredError, RandomInitialization>* model;
-    arma::mat input, target;
-    const size_t inSize;
-    const size_t outSize;
-    const size_t nPoints;
-    const size_t batchSize;
-  } function;
-
-  REQUIRE(CheckGradient(function) <= 1e-7);
-}
-
 // /**
 //  * Simple noisy linear module test.
 //  */
@@ -682,158 +395,6 @@ TEST_CASE("GradientLinear3DLayerTest", "[ANNLayerTest]")
 
 //   REQUIRE(CheckGradient(function) <= 1e-4);
 // }
-
-/**
- * Simple linear no bias module test.
- */
-TEST_CASE("SimpleLinearNoBiasLayerTest", "[ANNLayerTest]")
-{
-  arma::mat output, input, delta;
-  LinearNoBias module(10);
-  arma::mat weights(10 * 10, 1);
-  module.InputDimensions() = std::vector<size_t>({ 10 });
-  module.ComputeOutputDimensions();
-  module.SetWeights(weights.memptr());
-
-  module.Parameters().randu();
-
-  // Test the Forward function.
-  input = arma::zeros(10, 1);
-  module.Forward(input, output);
-  REQUIRE(0 == arma::accu(output));
-
-  // Test the Backward function.
-  module.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 0);
-}
-
-/**
- * Simple padding layer test.
- */
-TEST_CASE("SimplePaddingLayerTest", "[ANNLayerTest]")
-{
-  arma::mat output, input, delta;
-  Padding module(1, 2, 3, 4);
-  module.InputDimensions() = std::vector<size_t>({ 2, 5 });
-  module.ComputeOutputDimensions();
-
-  // Test the Forward function.
-  input = arma::randu(10, 1);
-  size_t totalOutputDimensions = module.OutputDimensions()[0];
-  for (size_t i = 1; i < module.OutputDimensions().size(); ++i)
-    totalOutputDimensions *= module.OutputDimensions()[i];
-  output.set_size(totalOutputDimensions, input.n_cols);
-  output.randu();
-  module.Forward(input, output);
-  REQUIRE(arma::accu(input) == Approx(arma::accu(output)));
-  REQUIRE(output.n_rows == (9 * 8)); // 2x5 --> 9x8
-
-  // Test the Backward function.
-  delta.set_size(input.n_rows, input.n_cols);
-  module.Backward(input, output, delta);
-  CheckMatrices(delta, input);
-
-  // Test forward function for multiple filters.
-  // Here it's 3 filters with height = 224, width = 224
-  // the output should be [226 * 226 * 3, 1] with 1 padding.
-  module = Padding(1, 1, 1, 1);
-  module.InputDimensions() = std::vector<size_t>({ 224, 224, 3 });
-  module.ComputeOutputDimensions();
-
-  input = arma::randu(224 * 224 * 3, 1);
-  totalOutputDimensions = module.OutputDimensions()[0];
-  for (size_t i = 1; i < module.OutputDimensions().size(); ++i)
-    totalOutputDimensions *= module.OutputDimensions()[i];
-  output.set_size(totalOutputDimensions, input.n_cols);
-  output.randu();
-  module.Forward(input, output);
-  REQUIRE(arma::accu(input) == Approx(arma::accu(output)));
-  REQUIRE(output.n_rows == (226 * 226 * 3));
-  REQUIRE(output.n_cols == 1);
-
-  // Test forward function for multiple batches with multiple filters.
-  // Here it's 3 filters with height = 244, width = 244
-  // the output should be [246 * 246 * 3, 3] with 1 padding.
-  module.InputDimensions() = std::vector<size_t>({ 244, 244, 3 });
-  module.ComputeOutputDimensions();
-  totalOutputDimensions = module.OutputDimensions()[0];
-  for (size_t i = 1; i < module.OutputDimensions().size(); ++i)
-    totalOutputDimensions *= module.OutputDimensions()[i];
-
-  input = arma::randu(244 * 244 * 3, 3);
-  output.set_size(totalOutputDimensions, input.n_cols);
-  output.randu();
-  module.Forward(input, output);
-  REQUIRE(output.n_rows == (246 * 246 * 3));
-  REQUIRE(output.n_cols == 3);
-  REQUIRE(arma::accu(input) == Approx(arma::accu(output)));
-}
-
-/**
- * Jacobian linear no bias module test.
- */
-TEST_CASE("JacobianLinearNoBiasLayerTest", "[ANNLayerTest]")
-{
-  for (size_t i = 0; i < 5; ++i)
-  {
-    const size_t inputElements = math::RandInt(2, 1000);
-    const size_t outputElements = math::RandInt(2, 1000);
-
-    arma::mat input;
-    input.set_size(inputElements, 1);
-
-    LinearNoBias module(outputElements);
-    arma::mat weights(inputElements * outputElements, 1);
-    module.InputDimensions() = std::vector<size_t>({ inputElements });
-    module.ComputeOutputDimensions();
-    module.SetWeights(weights.memptr());
-
-    module.Parameters().randu();
-
-    double error = JacobianTest(module, input);
-    REQUIRE(error <= 1e-5);
-  }
-}
-
-/**
- * LinearNoBias layer numerical gradient test.
- */
-TEST_CASE("GradientLinearNoBiasLayerTest", "[ANNLayerTest]")
-{
-  // LinearNoBias function gradient instantiation.
-  struct GradientFunction
-  {
-    GradientFunction() :
-        input(arma::randu(10, 1)),
-        target(arma::mat("0"))
-    {
-      model = new FFN<NegativeLogLikelihood, NguyenWidrowInitialization>();
-      model->ResetData(input, target);
-      model->Add<Linear>(10);
-      model->Add<LinearNoBias>(2);
-      model->Add<LogSoftMax>();
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 1);
-      model->Gradient(model->Parameters(), 0, gradient, 1);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<NegativeLogLikelihood, NguyenWidrowInitialization>* model;
-    arma::mat input, target;
-  } function;
-
-  REQUIRE(CheckGradient(function) <= 1e-4);
-}
 
 // /**
 //  * Jacobian negative log likelihood module test.
@@ -939,56 +500,6 @@ TEST_CASE("GradientFlexibleReLULayerTest", "[ANNLayerTest]")
   REQUIRE(CheckGradient(function) <= 1e-4);
 }
 */
-
-/**
- * Simple ELU module test.
- */
-TEST_CASE("SimpleELULayerTest", "[ANNLayerTest]")
-{
-  arma::mat input;
-  ELU module(0.5);
-  module.Training() = true;
-  module.InputDimensions() = std::vector<size_t>({ 1 });
-  module.ComputeOutputDimensions();
-  
-  // Test the Forward function.
-  input = arma::mat("-1 2");
-  arma::mat output(arma::size(input));
-  module.Forward(input, output);
-  REQUIRE(arma::accu(arma::abs(arma::mat("-0.316060 2") - output)) ==
-      Approx(0.0).margin(1e-4));
-
-  // Test the Backward function.
-  arma::mat delta(arma::size(input));
-  module.Backward(input, output, delta);
-  REQUIRE(arma::accu(arma::abs(arma::mat("-0.058136 2") - delta)) ==
-      Approx(0.0).margin(1e-4));
-}
-
-/**
- * Simple CELU module test.
- */
-TEST_CASE("SimpleCELULayerTest", "[ANNLayerTest]")
-{
-  arma::mat input;
-  CELU module(0.5);
-  module.Training() = true;
-  module.InputDimensions() = std::vector<size_t>({ 1 });
-  module.ComputeOutputDimensions();
-
-  // Test the Forward function.
-  input = arma::mat("-1 2");
-  arma::mat output(arma::size(input));
-  module.Forward(input, output);
-  REQUIRE(arma::accu(arma::abs(arma::mat("-0.432332 2") - output)) ==
-      Approx(0.0).margin(1e-4));
-
-  // Test the Backward function.
-  arma::mat delta(arma::size(input));
-  module.Backward(input, output, delta);
-  REQUIRE(arma::accu(arma::abs(arma::mat("-0.0585098 2") - delta)) ==
-      Approx(0.0).margin(1e-4));
-}
 
 /**
  * Jacobian MultiplyConstant module test.
@@ -2739,29 +2250,6 @@ TEST_CASE("LookupLayerParametersTest", "[ANNLayerTest]")
 */
 
 /**
- * Simple LogSoftMax module test.
- */
-TEST_CASE("SimpleLogSoftmaxLayerTest", "[ANNLayerTest]")
-{
-  arma::mat output, input, error, delta;
-  LogSoftMax module;
-
-  // Test the Forward function.
-  input = arma::mat("0.5; 0.5");
-  module.Forward(input, output);
-  REQUIRE(arma::accu(arma::abs(arma::mat("-0.6931; -0.6931") - output)) ==
-      Approx(0.0).margin(1e-3));
-
-  // Test the Backward function.
-  error = arma::zeros(input.n_rows, input.n_cols);
-  // Assume LogSoftmax layer is always associated with NLL output layer.
-  error(1, 0) = -1;
-  module.Backward(input, error, delta);
-  REQUIRE(arma::accu(arma::abs(arma::mat("1.6487; 0.6487") - delta)) ==
-      Approx(0.0).margin(1e-3));
-}
-
-/**
  * Simple Softmax module test.
  */
 TEST_CASE("SimpleSoftmaxLayerTest", "[ANNLayerTest]")
@@ -3054,155 +2542,6 @@ TEST_CASE("SimpleBicubicInterpolationLayerTest", "[ANNLayerTest]")
   CheckMatrices(unzoomedOutput1, expectedUnzoomed, 1e-6);
 }
 */
-
-/**
- * Tests the BatchNorm Layer, compares the layers parameters with
- * the values from another implementation.
- * Link to the implementation - http://cthorey.github.io./backpropagation/
- */
-TEST_CASE("BatchNormTest", "[ANNLayerTest]")
-{
-  arma::mat input, output;
-  input << 5.1 << 3.5 << 1.4 << arma::endr
-        << 4.9 << 3.0 << 1.4 << arma::endr
-        << 4.7 << 3.2 << 1.3 << arma::endr;
-
-  input = input.t();
-
-  // BatchNorm layer with average parameter set to true.
-  BatchNorm module1;
-  module1.Training() = true;
-  module1.InputDimensions() = std::vector<size_t>({ 3, 3 });
-  module1.ComputeOutputDimensions();
-  arma::mat moduleParams(module1.WeightSize(), 1);
-  module1.CustomInitialize(moduleParams, module1.WeightSize());
-  module1.SetWeights((double*) moduleParams.memptr());
-
-  // BatchNorm layer with average parameter set to false (using momentum).
-  BatchNorm module2(2, 2, 1e-5, false);
-  module2.Training() = true;
-  module2.InputDimensions() = std::vector<size_t>({ 3, 3 });
-  module2.ComputeOutputDimensions();
-  arma::mat moduleParams2(module2.WeightSize(), 1);
-  module2.CustomInitialize(moduleParams2, module2.WeightSize());
-  module2.SetWeights((double*) moduleParams2.memptr());
-
-  // Training Forward Pass Test.
-  output.set_size(module1.OutputSize(), 1);
-  input.reshape(9, 1);
-  module1.Forward(input, output);
-
- // Value calculates using torch.nn.BatchNorm1d(momentum = None).
-  arma::mat result;
-  output.reshape(3, 3);
-  result = { { 1.1658, 0.1100, -1.2758 },
-            { 1.2579, -0.0699, -1.1880},
-            { 1.1737, 0.0958, -1.2695 } };
-
-  CheckMatrices(output, result.t(), 1e-1);
-
-  output.set_size(module2.OutputSize(), 1);
-  module2.Forward(input, output);
-  output.reshape(3, 3);
-  CheckMatrices(output, result.t(), 1e-1);
-  result.clear();
-  output.clear();
-
- // Values calculated using torch.nn.BatchNorm1d(momentum = None).
-  output = module1.TrainingMean();
-  result = arma::mat({ 3.33333333, 3.1, 3.06666666 }).t();
-
-  CheckMatrices(output, result, 1e-1);
-
- // Values calculated using torch.nn.BatchNorm1d().
-  output = module2.TrainingMean();
-  result = arma::mat({ 0.3333, 0.3100, 0.3067 }).t();
-
-  CheckMatrices(output, result, 1e-1);
-  result.clear();
-
-  // Values calculated using torch.nn.BatchNorm1d(momentum = None).
-  output = module1.TrainingVariance();
-  result = arma::mat({ 3.4433, 3.0700, 2.9033 }).t();
-
-  CheckMatrices(output, result, 1e-1);
-  result.clear();
-
-  // Values calculated using torch.nn.BatchNorm1d().
-  output = module2.TrainingVariance();
-  result = arma::mat({ 1.2443, 1.2070, 1.1903 }).t();
-
-  CheckMatrices(output, result, 1e-1);
-  result.clear();
-
-  // Deterministic Forward Pass test.
-  module1.Training() = false;
-  output.set_size(module1.OutputSize(), 1);
-  module1.Forward(input, output);
-  output.reshape(3, 3);
-
-  // Values calculated using torch.nn.BatchNorm1d(momentum = None).
-  result = { { 0.9521, 0.0898, -1.0419 },
-            { 1.0273, -0.0571, -0.9702 },
-            { 0.9586, 0.0783, -1.0368 } };
-
-  CheckMatrices(output, result.t(), 1e-1);
-
-  // Values calculated using torch.nn.BatchNorm1d().
-  module2.Training() = false;
-  output.set_size(module2.OutputSize(), 1);
-  module2.Forward(input, output);
-  output.reshape(3, 3);
-
-  result = { { 4.2731, 2.8388, 0.9562 },
-             { 4.1779, 2.4485, 0.9921 },
-             { 4.0268, 2.6519, 0.9105 } };
-
-  CheckMatrices(output, result.t(), 1e-1);
-}
-
-/**
- * BatchNorm layer numerical gradient test.
- */
-TEST_CASE("GradientBatchNormTest", "[ANNLayerTest]")
-{
-  // Add function gradient instantiation.
-  struct GradientFunction
-  {
-    GradientFunction() :
-        input(arma::randn(32, 2048)),
-        target(arma::zeros(1, 2048))
-    {
-      model = new FFN<NegativeLogLikelihood, NguyenWidrowInitialization>();
-      model->ResetData(input, target);
-      model->Add<Linear>(4);
-      model->Add<BatchNorm>();
-      model->Add<Linear>(2);
-      model->Add<LogSoftMax>();
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 2048);
-      model->Gradient(model->Parameters(), 0, gradient, 2048);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<NegativeLogLikelihood, NguyenWidrowInitialization>* model;
-    arma::mat input, target;
-  } function;
-
-  double gradient = CheckGradient(function);
-
-  REQUIRE(gradient < 1e-1);
-}
 
 /**
  * VirtualBatchNorm layer numerical gradient test.
@@ -4572,52 +3911,6 @@ TEST_CASE("HighwayLayerParametersTest", "[ANNLayerTest]")
 //   REQUIRE(arma::accu(delta) == 0);
 // }
 
-// General ANN serialization test.
-template<typename LayerType>
-void ANNLayerSerializationTest(LayerType& layer)
-{
-  arma::mat input(5, 100, arma::fill::randu);
-  arma::mat output(5, 100, arma::fill::randu);
-
-  FFN<> model;
-  model.Add<Linear>(10);
-  model.Add<LayerType>(layer);
-  model.Add<ReLU>();
-  model.Add<Linear>(output.n_rows);
-  model.Add<LogSoftMax>();
-
-  ens::StandardSGD opt(0.1, 1, 5, -100, false);
-  model.Train(input, output, opt);
-
-  arma::mat originalOutput;
-  model.Predict(input, originalOutput);
-
-  // Now serialize the model.
-  FFN<> xmlModel, jsonModel, binaryModel;
-  SerializeObjectAll(model, xmlModel, jsonModel, binaryModel);
-
-  // Ensure that predictions are the same.
-  arma::mat modelOutput, xmlOutput, jsonOutput, binaryOutput;
-  model.Predict(input, modelOutput);
-  xmlModel.Predict(input, xmlOutput);
-  jsonModel.Predict(input, jsonOutput);
-  binaryModel.Predict(input, binaryOutput);
-
-  CheckMatrices(originalOutput, modelOutput, 1e-5);
-  CheckMatrices(originalOutput, xmlOutput, 1e-5);
-  CheckMatrices(originalOutput, jsonOutput, 1e-5);
-  CheckMatrices(originalOutput, binaryOutput, 1e-5);
-}
-
-/**
- * Simple serialization test for batch normalization layer.
- */
-TEST_CASE("BatchNormSerializationTest", "[ANNLayerTest]")
-{
-  BatchNorm layer;
-  ANNLayerSerializationTest(layer);
-}
-
 // /**
 //  * Simple serialization test for layer normalization layer.
 //  */
@@ -4626,144 +3919,6 @@ TEST_CASE("BatchNormSerializationTest", "[ANNLayerTest]")
 //   LayerNorm<> layer(10);
 //   ANNLayerSerializationTest(layer);
 // }
-
-/**
- * Test that the functions that can modify and access the parameters of the
- * Convolution layer work.
- */
-TEST_CASE("ConvolutionLayerParametersTest", "[ANNLayerTest]")
-{
-  // Parameter order: outSize, kW, kH, dW, dH, padW, padH, paddingType.
-  Convolution layer1(2, 3, 4, 5, 6, std::tuple<size_t, size_t>(7, 8),
-      std::tuple<size_t, size_t>(9, 10), "none");
-  Convolution layer2(3, 4, 5, 6, 7, std::tuple<size_t, size_t>(8, 9),
-      std::tuple<size_t, size_t>(10, 11), "none");
-
-  // Make sure we can get the parameters successfully.
-  REQUIRE(layer1.KernelWidth() == 3);
-  REQUIRE(layer1.KernelHeight() == 4);
-  REQUIRE(layer1.StrideWidth() == 5);
-  REQUIRE(layer1.StrideHeight() == 6);
-  REQUIRE(layer1.PadWLeft() == 7);
-  REQUIRE(layer1.PadWRight() == 8);
-  REQUIRE(layer1.PadHTop() == 9);
-  REQUIRE(layer1.PadHBottom() == 10);
-
-  // Now modify the parameters to match the second layer.
-  layer1.KernelWidth() = 4;
-  layer1.KernelHeight() = 5;
-  layer1.StrideWidth() = 6;
-  layer1.StrideHeight() = 7;
-  layer1.PadWLeft() = 8;
-  layer1.PadWRight() = 9;
-  layer1.PadHTop() = 10;
-  layer1.PadHBottom() = 11;
-
-  // Now ensure all results are the same.
-  REQUIRE(layer1.KernelWidth() == layer2.KernelWidth());
-  REQUIRE(layer1.KernelHeight() == layer2.KernelHeight());
-  REQUIRE(layer1.StrideWidth() == layer2.StrideWidth());
-  REQUIRE(layer1.StrideHeight() == layer2.StrideHeight());
-  REQUIRE(layer1.PadWLeft() == layer2.PadWLeft());
-  REQUIRE(layer1.PadWRight() == layer2.PadWRight());
-  REQUIRE(layer1.PadHTop() == layer2.PadHTop());
-  REQUIRE(layer1.PadHBottom() == layer2.PadHBottom());
-}
-
-/**
- * Test that the padding options are working correctly in Convolution layer.
- */
-TEST_CASE("ConvolutionLayerPaddingTest", "[ANNLayerTest]")
-{
-  arma::mat output, input, delta;
-
-  // Check valid padding option.
-  Convolution module1(1, 3, 3, 1, 1, std::tuple<size_t, size_t>(1, 1),
-      std::tuple<size_t, size_t>(1, 1), "valid");
-  module1.InputDimensions() = std::vector<size_t>({ 7, 7 });
-  module1.ComputeOutputDimensions();
-  arma::mat weights1(module1.WeightSize(), 1);
-  REQUIRE(weights1.n_elem == 10);
-  module1.SetWeights(weights1.memptr());
-
-  // Test the Forward function.
-  input = arma::linspace<arma::colvec>(0, 48, 49);
-  output.set_size(module1.OutputSize(), 1);
-  module1.Parameters().zeros();
-  module1.Forward(input, output);
-
-  REQUIRE(arma::accu(output) == 0);
-  REQUIRE(output.n_rows == 25);
-  REQUIRE(output.n_cols == 1);
-
-  // Test the Backward function.
-  delta.set_size(arma::size(input));
-  module1.Backward(input, output, delta);
-
-  // Check same padding option.
-  Convolution module2(1, 3, 3, 1, 1, std::tuple<size_t, size_t>(0, 0),
-      std::tuple<size_t, size_t>(0, 0), "same");
-  module2.InputDimensions() = std::vector<size_t>({ 7, 7 });
-  module2.ComputeOutputDimensions();
-  arma::mat weights2(module2.WeightSize(), 1);
-  REQUIRE(weights2.n_elem == 10);
-  module2.SetWeights(weights2.memptr());
-
-  // Test the forward function.
-  input = arma::linspace<arma::colvec>(0, 48, 49);
-  output.set_size(module2.OutputSize(), 1);
-  module2.Parameters().zeros();
-  module2.Forward(input, output);
-
-  REQUIRE(arma::accu(output) == 0);
-  REQUIRE(output.n_rows == 49);
-  REQUIRE(output.n_cols == 1);
-
-  // Test the backward function.
-  delta.set_size(arma::size(input));
-  module2.Backward(input, output, delta);
-}
-
-/**
- * Convolution layer numerical gradient test.
- */
-TEST_CASE("GradientConvolutionLayerTest", "[ANNLayerTest]")
-{
-  struct GradientFunction
-  {
-    GradientFunction() :
-        input(arma::linspace<arma::colvec>(0, 35, 36)),
-        target(arma::mat("1"))
-    {
-      model = new FFN<NegativeLogLikelihood, RandomInitialization>();
-      model->ResetData(input, target);
-      model->Add<Convolution>(1, 3, 3, 1, 1, std::tuple<size_t, size_t>(0, 0),
-          std::tuple<size_t, size_t>(0, 0), "same");
-      model->Add<LogSoftMax>();
-
-      model->InputDimensions() = std::vector<size_t>({ 6, 6 });
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 1);
-      model->Gradient(model->Parameters(), 0, gradient, 1);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<NegativeLogLikelihood, RandomInitialization>* model;
-    arma::mat input, target;
-  } function;
-
-  REQUIRE(CheckGradient(function) < 1e3);
-}
 
 /**
  * Test that the padding options in Transposed Convolution layer.
@@ -5063,166 +4218,6 @@ TEST_CASE("IdentityTestCase", "[ANNLayerTest]")
 }
 
 /**
- * Simple test for Mean Pooling layer.
- */
-TEST_CASE("MeanPoolingTestCase", "[ANNLayerTest]")
-{
-  // For rectangular input to pooling layers.
-  arma::mat input = arma::mat(28, 1);
-  input.zeros();
-  input(0) = input(16) = 1;
-  input(1) = input(17) = 2;
-  input(2) = input(18) = 3;
-  input(3) = input(19) = 4;
-  input(4) = input(20) = 5;
-  input(5) = input(23) = 6;
-  input(6) = input(24) = 7;
-  input(14) = input(25) = 8;
-  input(15) = input(26) = 9;
-
-  MeanPooling module1(2, 2, 2, 2, false);
-  MeanPooling module2(2, 2, 2, 2, true);
-  module1.InputDimensions() = std::vector<size_t>({ 7, 4 });
-  module1.ComputeOutputDimensions();
-  module2.InputDimensions() = std::vector<size_t>({ 7, 4 });
-  module2.ComputeOutputDimensions();
-
-  // Calculated using torch.nn.MeanPool2d().
-  arma::mat result1 = { { 0.7500, 4.2500 },
-                        { 1.7500, 4.0000 },
-                        { 2.7500, 6.0000 },
-                        { 3.5000, 2.5000 } };
-  
-  arma::mat result2 = { { 0.7500, 4.2500 },
-                        { 1.7500, 4.0000 },
-                        { 2.7500, 6.0000 } };
-
-  arma::mat output1, output2;
-  output1.set_size(8, 1);
-  output2.set_size(6, 1);
-  module1.Forward(input, output1);
-  REQUIRE(arma::accu(output1) == 25.5);
-  module2.Forward(input, output2);
-  REQUIRE(arma::accu(output2) == 19.5);
-  output1.reshape(4, 2);
-  output2.reshape(3, 2);
-  CheckMatrices(output1, result1, 1e-1);
-  CheckMatrices(output2, result2, 1e-1);
-
-  arma::mat prevDelta1 = { { 3.6000, -0.9000 },
-                           { 3.6000, -0.9000 },
-                           { 3.6000, -0.9000 },
-                           { 3.6000, -0.9000 } };
-
-  arma::mat prevDelta2 = { { 3.6000, -0.9000 }, 
-                           { 3.6000, -0.9000 }, 
-                           { 3.6000, -0.9000 } };
-
-  arma::mat delta1, delta2;
-  delta1.set_size(28, 1);
-  delta2.set_size(28, 1);
-  prevDelta1.reshape(8, 1);
-  prevDelta2.reshape(6, 1);
-  module1.Backward(input, prevDelta1, delta1);
-  REQUIRE(arma::accu(delta1) == Approx(10.8).epsilon(1e-3));
-  module2.Backward(input, prevDelta2, delta2);
-  REQUIRE(arma::accu(delta2) == Approx(8.1).epsilon(1e-3));
-}
-
-/**
- * Simple test for Max Pooling layer.
- */
-TEST_CASE("MaxPoolingTestCase", "[ANNLayerTest]")
-{
-  // For rectangular input to pooling layers.
-  arma::mat input = arma::mat(12, 1);
-  arma::mat output;
-  input.zeros();
-  input(0) = 1;
-  input(1) = 2;
-  input(2) = 3;
-  input(3) = input(8) = 7;
-  input(4) = 4;
-  input(5) = 5;
-  input(6) = input(7) = 6;
-  input(10) = 8;
-  input(11) = 9;
-  // Output-Size should be 2 x 2.
-  output.set_size(4, 1);
-
-  // Square output.
-  MaxPooling module1(2, 2, 2, 1);
-  module1.InputDimensions() = std::vector<size_t>({ 4, 3 });
-  module1.ComputeOutputDimensions();
-  module1.Forward(input, output);
-  // Calculated using torch.nn.MaxPool2d().
-  REQUIRE(arma::accu(output) == 28);
-  REQUIRE(output.n_elem == 4);
-  REQUIRE(output.n_cols == 1);
-
-  // For Square input.
-  input = arma::mat(9, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(3) = 3;
-  input(6) = 3;
-  // Output-Size should be 1 x 2.
-  output.set_size(2, 1);
-
-  // Rectangular output.
-  MaxPooling module2(3, 2, 3, 1);
-  module2.InputDimensions() = std::vector<size_t>({ 3, 3 });
-  module2.ComputeOutputDimensions();
-  module2.Forward(input, output);
-  // Calculated using torch.nn.MaxPool2d().
-  REQUIRE(arma::accu(output) == 12.0);
-  REQUIRE(output.n_elem == 2);
-  REQUIRE(output.n_cols == 1);
-
-  // For Square input.
-  input = arma::mat(16, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(4) = 3;
-  input(8) = 3;
-  // Output-Size should be 3 x 3.
-  output.set_size(9, 1);
-
-  // Square output.
-  MaxPooling module3(2, 2, 1, 1);
-  module3.InputDimensions() = std::vector<size_t>({ 4, 4 });
-  module3.ComputeOutputDimensions();
-  module3.Forward(input, output);
-  // Calculated using torch.nn.MaxPool2d().
-  REQUIRE(arma::accu(output) == 30.0);
-  REQUIRE(output.n_elem == 9);
-  REQUIRE(output.n_cols == 1);
-
-  // For Rectangular input.
-  input = arma::mat(6, 1);
-  input.zeros();
-  input(0) = 1;
-  input(1) = 1;
-  input(3) = 1;
-  // Output-Size should be 2 x 2.
-  output.set_size(4, 1);
-
-  // Square output.
-  MaxPooling module4(2, 1, 1, 1);
-  module4.InputDimensions() = std::vector<size_t>({ 3, 2 });
-  module4.ComputeOutputDimensions();
-  module4.Forward(input, output);
-  // Calculated using torch.nn.MaxPool2d().
-  REQUIRE(arma::accu(output) == 3);
-  REQUIRE(output.n_elem == 4);
-  REQUIRE(output.n_cols == 1);
-}
-
-/**
  * Test that the functions that can modify and access the parameters of the
  * Glimpse layer work.
  *
@@ -5268,223 +4263,6 @@ TEST_CASE("ReinforceNormalLayerParametersTest", "[ANNLayerTest]")
 }
 */
 
-/**
- * Simple test for Adaptive pooling for Max Pooling layer.
- */
-TEST_CASE("AdaptiveMaxPoolingTestCase", "[ANNLayerTest]")
-{
-  // For rectangular input.
-  arma::mat input = arma::mat(12, 1);
-  arma::mat output, delta;
-
-  input.zeros();
-  input(0) = 1;
-  input(1) = 2;
-  input(2) = 3;
-  input(3) = input(8) = 7;
-  input(4) = 4;
-  input(5) = 5;
-  input(6) = input(7) = 6;
-  input(10) = 8;
-  input(11) = 9;
-  // Output-Size should be 2 x 2.
-  // Square output.
-  AdaptiveMaxPooling module1(2, 2);
-  output.set_size(4, 1);
-  module1.InputDimensions() = std::vector<size_t>({ 4, 3 });
-  module1.ComputeOutputDimensions();
-  module1.Training() = true;
-  module1.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveMaxPool2d().
-  REQUIRE(arma::accu(output) == 28);
-  REQUIRE(output.n_elem == 4);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(12, 1);
-  module1.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 28.0);
-
-  // For Square input.
-  input = arma::mat(9, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(3) = 3;
-  input(6) = 3;
-  // Output-Size should be 1 x 2.
-  // Rectangular output.
-  AdaptiveMaxPooling module2(2, 1);
-  output.set_size(2, 1);
-  module2.InputDimensions() = std::vector<size_t>({ 3, 3 });
-  module2.ComputeOutputDimensions();
-  module2.Training() = true;
-  module2.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveMaxPool2d().
-  REQUIRE(arma::accu(output) == 15.0);
-  REQUIRE(output.n_elem == 2);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(9, 1);
-  module2.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 15.0);
-
-  // For Square input.
-  input = arma::mat(16, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(4) = 3;
-  input(8) = 3;
-  // Output-Size should be 3 x 3.
-  // Square output.
-  AdaptiveMaxPooling module3(3, 3);
-  output.set_size(9, 1);
-  module3.InputDimensions() = std::vector<size_t>({ 4, 4 });
-  module3.ComputeOutputDimensions();
-  module3.Training() = true;
-  module3.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveMaxPool2d().
-  REQUIRE(arma::accu(output) == 30.0);
-  REQUIRE(output.n_elem == 9);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(16, 1);
-  module3.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 30.0);
-
-  // For Rectangular input.
-  input = arma::mat(20, 1);
-  input.zeros();
-  input(0) = 1;
-  input(1) = 1;
-  input(3) = 1;
-  // Output-Size should be 2 x 2.
-  // Square output.
-  AdaptiveMaxPooling module4(2, 2);
-  output.set_size(4, 1);
-  module4.InputDimensions() = std::vector<size_t>({ 5, 4 });
-  module4.ComputeOutputDimensions();
-  module4.Training() = true;
-  module4.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveMaxPool2d().
-  REQUIRE(arma::accu(output) == 2);
-  REQUIRE(output.n_elem == 4);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(20, 1);
-  module4.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 2.0);
-}
-
-/**
- * Simple test for Adaptive pooling for Mean Pooling layer.
- *
- */
-TEST_CASE("AdaptiveMeanPoolingTestCase", "[ANNLayerTest]")
-{
-  // For rectangular input.
-  arma::mat input = arma::mat(12, 1);
-  arma::mat output, delta;
-
-  input.zeros();
-  input(0) = 1;
-  input(1) = 2;
-  input(2) = 3;
-  input(3) = input(8) = 7;
-  input(4) = 4;
-  input(5) = 5;
-  input(6) = input(7) = 6;
-  input(10) = 8;
-  input(11) = 9;
-  // Output-Size should be 2 x 2.
-  // Square output.
-  AdaptiveMeanPooling module1(2, 2);
-  output.set_size(4, 1);
-  module1.InputDimensions() = std::vector<size_t>({ 4, 3 });
-  module1.ComputeOutputDimensions();
-  module1.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveAvgPool2d().
-  REQUIRE(arma::accu(output) == 19.75);
-  REQUIRE(output.n_elem == 4);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(12, 1);
-  module1.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 19.75);
-
-  // For Square input.
-  input = arma::mat(9, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(3) = 3;
-  input(6) = 3;
-  // Output-Size should be 1 x 2.
-  // Rectangular output.
-  AdaptiveMeanPooling module2(1, 2);
-  output.set_size(2, 1);
-  module2.InputDimensions() = std::vector<size_t>({ 3, 3 });
-  module2.ComputeOutputDimensions();
-  module2.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveAvgPool2d().
-  REQUIRE(arma::accu(output) == 4.5);
-  REQUIRE(output.n_elem == 2);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(9, 1);
-  module2.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 4.50);
-
-  // For Square input.
-  input = arma::mat(16, 1);
-  input.zeros();
-  input(0) = 6;
-  input(1) = 3;
-  input(2) = 9;
-  input(4) = 3;
-  input(8) = 3;
-  // Output-Size should be 3 x 3.
-  // Square output.
-  AdaptiveMeanPooling module3(3, 3);
-  output.set_size(9, 1);
-  module3.InputDimensions() = std::vector<size_t>({ 4, 4 });
-  module3.ComputeOutputDimensions();
-  module3.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveAvgPool2d().
-  REQUIRE(arma::accu(output) == 10.5);
-  REQUIRE(output.n_elem == 9);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(16, 1);
-  module3.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 10.5);
-
-  // For Rectangular input.
-  input = arma::mat(24, 1);
-  input.zeros();
-  input(0) = 3;
-  input(1) = 3;
-  input(4) = 3;
-  // Output-Size should be 3 x 3.
-  // Square output.
-  AdaptiveMeanPooling module4(3, 3);
-  output.set_size(9, 1);
-  module4.InputDimensions() = std::vector<size_t>({ 6, 4 });
-  module4.ComputeOutputDimensions();
-  module4.Forward(input, output);
-  // Calculated using torch.nn.AdaptiveAvgPool2d().
-  REQUIRE(arma::accu(output) == 2.25);
-  REQUIRE(output.n_elem == 9);
-  REQUIRE(output.n_cols == 1);
-  // Test the Backward Function.
-  delta.set_size(24, 1);
-  module4.Backward(input, output, delta);
-  REQUIRE(arma::accu(delta) == 2.25);
-}
-
 /*
 TEST_CASE("TransposedConvolutionalLayerOptionalParameterTest", "[ANNLayerTest]")
 {
@@ -5500,219 +4278,6 @@ TEST_CASE("TransposedConvolutionalLayerOptionalParameterTest", "[ANNLayerTest]")
   delete decoder;
 }
 */
-
-TEST_CASE("BatchNormWithMinBatchesTest", "[ANNLayerTest]")
-{
-  arma::mat input, output, result, runningMean, runningVar, delta;
-
-  // The input test matrix is of the form 3 x 2 x 4 x 1 where
-  // number of images are 3 and number of feature maps are 2.
-  input = { { 1, 446, 42 },
-            { 2, 16, 63 },
-            { 3, 13, 63 },
-            { 4, 21, 21 },
-            { 1, 13, 11 },
-            { 32, 45, 42 },
-            { 22, 16, 63 },
-            { 32, 13, 42 } };
-
-  // Output calculated using torch.nn.BatchNorm2d().
-  result = { { -0.4786, 3.2634, -0.1338 },
-             { -0.4702, -0.3525, 0.0427 },
-             { -0.4618, -0.3777, 0.0427 },
-             { -0.4534, -0.3104, -0.3104 },
-             { -1.5429, -0.8486, -0.9643 },
-             { 0.2507, 1.0029, 0.8293 },
-             { -0.3279, -0.675, 2.0443 },
-             { 0.2507 , -0.8486 , 0.8293 } };
-
-  // Check correctness of batch normalization.
-  BatchNorm module1(2, 2, 1e-5, false, 0.1);
-  module1.Training() = true;
-  module1.InputDimensions() = std::vector<size_t>({ 1, 4, 2 });
-  module1.ComputeOutputDimensions();
-  arma::mat moduleParams(module1.WeightSize(), 1);
-  module1.CustomInitialize(moduleParams, module1.WeightSize());
-  module1.SetWeights((double*) moduleParams.memptr());
-  output.set_size(8, 3);
-  module1.Forward(input, output);
-  CheckMatrices(output, result, 1e-1);
-
-  // Check values for running mean and running variance.
-  // Calculated using torch.nn.BatchNorm2d().
-  runningMean = arma::mat(2, 1);
-  runningVar = arma::mat(2, 1);
-  runningMean(0) = 5.7917;
-  runningMean(1) = 2.76667;
-  runningVar(0) = 1543.6545;
-  runningVar(1) = 33.488;
-
-  CheckMatrices(runningMean, module1.TrainingMean(), 1e-3);
-  CheckMatrices(runningVar, module1.TrainingVariance(), 1e-2);
-
-  // Check correctness of layer when running mean and variance
-  // are updated using cumulative average.
-  BatchNorm module2;
-  module2.Training() = true;
-  module2.InputDimensions() = std::vector<size_t>({ 1, 4, 2 });
-  module2.ComputeOutputDimensions();
-  arma::mat moduleParams2(module2.WeightSize(), 1);
-  module2.CustomInitialize(moduleParams2, module2.WeightSize());
-  module2.SetWeights((double*) moduleParams2.memptr());
-  output.set_size(8, 3);
-  module2.Forward(input, output);
-  CheckMatrices(output, result, 1e-1);
-
-  // Check values for running mean and running variance.
-  // Calculated using torch.nn.BatchNorm2d().
-  runningMean(0) = 57.9167;
-  runningMean(1) = 27.6667;
-  runningVar(0) = 15427.5380;
-  runningVar(1) = 325.8787;
-
-  CheckMatrices(runningMean, module2.TrainingMean(), 1e-2);
-  CheckMatrices(runningVar, module2.TrainingVariance(), 1e-2);
-
-  // Check correctness when model is testing.
-  arma::mat deterministicOutput;
-  module1.Training() = false;
-  deterministicOutput.set_size(8, 3);
-  module1.Forward(input, deterministicOutput);
-
-  result.clear();
-  result = { { -0.12195, 11.20426, 0.92158 },
-             { -0.0965, 0.259824, 1.4560 },
-             { -0.071054, 0.183567, 1.45607 },
-             { -0.045601, 0.3870852, 0.38708 },
-             { -0.305288, 1.7683, 1.4227 },
-             { 5.05166, 7.29812, 6.7797 },
-             { 3.323614, 2.2867, 10.4086 },
-             { 5.05166, 1.7683, 6.7797 } };
-
-  CheckMatrices(result, deterministicOutput, 1e-1);
-}
-
-/**
- * Batch Normalization layer numerical gradient test.
- */
-TEST_CASE("GradientBatchNormWithConvolutionTest", "[ANNLayerTest]")
-{
-  struct GradientFunction
-  {
-    GradientFunction() :
-        input(arma::randn(16, 1024)),
-        target(arma::zeros(1, 1024))
-    {
-      model = new FFN<NegativeLogLikelihood, NguyenWidrowInitialization>();
-      model->ResetData(input, target);
-      model->Add<Convolution>(2, 3, 3);
-      model->Add<BatchNorm>();
-      model->Add<Linear>(2);
-      model->Add<LogSoftMax>();
-
-      model->InputDimensions() = std::vector<size_t>({ 4, 4, 1 });
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 1024);
-      model->Gradient(model->Parameters(), 0, gradient, 1024);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<NegativeLogLikelihood, NguyenWidrowInitialization>* model;
-    arma::mat input, target;
-  } function;
-
-  double gradient = CheckGradient(function);
-
-  REQUIRE(gradient < 1e-1);
-}
-
-TEST_CASE("ConvolutionLayerTestCase", "[ANNLayerTest]")
-{
-  arma::mat input, output;
-
-  // The input test matrix is of the form 3 x 2 x 4 x 1 where
-  // number of images are 3 and number of feature maps are 2.
-  input = { { 1, 446, 42 },
-            { 2, 16, 63 },
-            { 3, 13, 63 },
-            { 4, 21, 21 },
-            { 1, 13, 11 },
-            { 32, 45, 42 },
-            { 22, 16 , 63 },
-            { 32, 13 , 42 } };
-
-  Convolution layer(4, 1, 1, 1, 1, 0, 0);
-  layer.InputDimensions() = std::vector<size_t>({ 4, 1, 2 });
-  layer.ComputeOutputDimensions();
-  arma::mat layerWeights(layer.WeightSize(), 1);
-  layer.SetWeights(layerWeights.memptr());
-  output.set_size(layer.OutputSize(), 3);
-
-  // Set weights to 1.0 and bias to 0.0.
-  layer.Weight().fill(1.0);
-  layer.Bias().zeros();
-  layer.Forward(input, output);
-
-  // Value calculated using torch.nn.Conv2d().
-  REQUIRE(arma::accu(output) == 4108);
-
-  // Set bias to one.
-  layer.Bias().fill(1.0);
-  layer.Forward(input, output);
-
-  // Value calculated using torch.nn.Conv2d().
-  REQUIRE(arma::accu(output) == 4156);
-}
-
-TEST_CASE("NoBiasConvolutionLayerTestCase", "[ANNLayerTest]")
-{
-  arma::mat input, output;
-
-  // The input test matrix is of the form 3 x 2 x 4 x 1 where
-  // number of images are 3 and number of feature maps are 2.
-  input = { { 1, 446, 42 },
-            { 2, 16, 63 },
-            { 3, 13, 63 },
-            { 4, 21, 21 },
-            { 1, 13, 11 },
-            { 32, 45, 42 },
-            { 22, 16 , 63 },
-            { 32, 13 , 42 } };
-
-  Convolution layer(4, 1, 1, 1, 1, 0, 0, "none", false);
-  layer.InputDimensions() = std::vector<size_t>({ 4, 1, 2 });
-  layer.ComputeOutputDimensions();
-  REQUIRE(layer.WeightSize() == 8);
-  arma::mat layerWeights(layer.WeightSize(), 1);
-  layer.SetWeights(layerWeights.memptr());
-  REQUIRE(layer.Bias().n_elem == 0);
-  output.set_size(layer.OutputSize(), 3);
-
-  // Set weights to 1.0 and bias to 0.0.
-  layer.Weight().fill(1.0);
-  layer.Bias().zeros();
-  layer.Forward(input, output);
-
-  // Value calculated using torch.nn.Conv2d().
-  REQUIRE(arma::accu(output) == 4108);
-
-  // Set bias to one.
-  layer.Bias().fill(1.0);
-  layer.Forward(input, output);
-
-  // Value calculated using torch.nn.Conv2d().
-  REQUIRE(arma::accu(output) == 4108);
-}
 
 // /**
 //  * Linear module weight initialization test.
@@ -5766,39 +4331,6 @@ TEST_CASE("NoBiasConvolutionLayerTestCase", "[ANNLayerTest]")
 //   REQUIRE(module.Parameters().n_rows
 //       == (outSize * inSize * kernelWidth * kernelHeight) + outSize);
 // }
-
-/**
- * Convolution module weight initialization test.
- */
-TEST_CASE("ConvolutionLayerWeightInitializationTest", "[ANNLayerTest]")
-{
-  size_t inSize = 2, outSize = 3;
-  size_t kernelWidth = 4, kernelHeight = 5;
-  Convolution module = Convolution(outSize,
-      kernelWidth, kernelHeight, 6, 7, std::tuple<size_t, size_t>(8, 9),
-      std::tuple<size_t, size_t>(10, 11), "none");
-  module.InputDimensions() = std::vector<size_t>({ 12, 13, 2 });
-  module.ComputeOutputDimensions();
-  arma::mat weights(module.WeightSize(), 1);
-  module.SetWeights(weights.memptr());
-
-  RandomInitialization().Initialize(module.Weight());
-  module.Bias().ones();
-
-  REQUIRE(std::equal(module.Weight().begin(),
-      module.Weight().end(), module.Parameters().begin()));
-
-  REQUIRE(std::equal(module.Bias().begin(),
-      module.Bias().end(), module.Parameters().end() - outSize));
-
-  REQUIRE(module.Weight().n_rows == kernelWidth);
-  REQUIRE(module.Weight().n_cols == kernelHeight);
-  REQUIRE(module.Weight().n_slices == outSize * inSize);
-  REQUIRE(module.Bias().n_rows == outSize);
-  REQUIRE(module.Bias().n_cols == 1);
-  REQUIRE(module.Parameters().n_rows
-      == (outSize * inSize * kernelWidth * kernelHeight) + outSize);
-}
 
 /**
  * Transposed Convolution module weight initialization test.
