+++ src/mlpack/tests/ann/not_adapted/dcgan_test.cpp
