+++ src/mlpack/tests/ann/reward_clipping_test.cpp
@@ -30,7 +30,7 @@
 
 #include <ensmallen.hpp>
 
-#include "catch.hpp"
+#include "../catch.hpp"
 
 using namespace mlpack;
 using namespace mlpack::ann;
