+++ src/mlpack/tests/string_encoding_test.cpp
@@ -77,7 +77,7 @@ void CheckVectors(const vector<vector<ValueType>>& a,
  */
 TEST_CASE("DictionaryEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   arma::mat output;
   DictionaryEncoding<SplitByAnyOf::TokenType> encoder;
@@ -114,7 +114,7 @@ TEST_CASE("DictionaryEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("UnicodeDictionaryEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   arma::mat output;
   DictionaryEncoding<SplitByAnyOf::TokenType> encoder;
@@ -148,7 +148,7 @@ TEST_CASE("UnicodeDictionaryEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("OnePassDictionaryEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   vector<vector<size_t>> output;
   DictionaryEncoding<SplitByAnyOf::TokenType> encoder(
@@ -185,10 +185,10 @@ TEST_CASE("OnePassDictionaryEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("SplitByAnyOfTokenizerTest", "[StringEncodingTest]")
 {
-  std::vector<STRING_VIEW> tokens;
-  STRING_VIEW line(stringEncodingInput[0]);
+  std::vector<MLPACK_STRING_VIEW> tokens;
+  MLPACK_STRING_VIEW line(stringEncodingInput[0]);
   SplitByAnyOf tokenizer(" ,.");
-  STRING_VIEW token = tokenizer(line);
+  MLPACK_STRING_VIEW token = tokenizer(line);
 
   while (!token.empty())
   {
@@ -223,10 +223,10 @@ TEST_CASE("SplitByAnyOfTokenizerUnicodeTest", "[StringEncodingTest]")
     "\xE2\x93\x82\xE2\x93\x81\xE2\x93\x85\xE2\x92\xB6\xE2\x92\xB8\xE2\x93\x80"
   };
 
-  std::vector<STRING_VIEW> tokens;
-  STRING_VIEW line(stringEncodingUtf8Input[2]);
+  std::vector<MLPACK_STRING_VIEW> tokens;
+  MLPACK_STRING_VIEW line(stringEncodingUtf8Input[2]);
   SplitByAnyOf tokenizer(" ,.");
-  STRING_VIEW token = tokenizer(line);
+  MLPACK_STRING_VIEW token = tokenizer(line);
   while (!token.empty())
   {
     tokens.push_back(token);
@@ -295,7 +295,7 @@ TEST_CASE("OnePassDictionaryEncodingIndividualCharactersTest",
  */
 TEST_CASE("StringEncodingCopyTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
   arma::sp_mat output;
   DictionaryEncoding<SplitByAnyOf::TokenType> encoderCopy;
   SplitByAnyOf tokenizer(" ,.");
@@ -331,7 +331,7 @@ TEST_CASE("StringEncodingCopyTest", "[StringEncodingTest]")
  */
 TEST_CASE("StringEncodingMoveTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
   arma::sp_mat output;
   DictionaryEncoding<SplitByAnyOf::TokenType> encoderCopy;
   SplitByAnyOf tokenizer(" ,.");
@@ -390,21 +390,21 @@ void CheckDictionaries(const StringEncodingDictionary<TokenType>& expected,
 
 /**
  * This is a specialization of the CheckDictionaries() function for
- * the STRING_VIEW token type.
+ * the MLPACK_STRING_VIEW token type.
  */
 template<>
 void CheckDictionaries(
-    const StringEncodingDictionary<STRING_VIEW>& expected,
-    const StringEncodingDictionary<STRING_VIEW>& obtained)
+    const StringEncodingDictionary<MLPACK_STRING_VIEW>& expected,
+    const StringEncodingDictionary<MLPACK_STRING_VIEW>& obtained)
 {
   /* MapType is equal to
    *
-   * std::unordered_map<STRING_VIEW,
+   * std::unordered_map<MLPACK_STRING_VIEW,
    *                    size_t,
-   *                    boost::hash<STRING_VIEW>>.
+   *                    boost::hash<MLPACK_STRING_VIEW>>.
    */
   using MapType =
-      typename StringEncodingDictionary<STRING_VIEW>::MapType;
+      typename StringEncodingDictionary<MLPACK_STRING_VIEW>::MapType;
 
   const std::deque<std::string>& expectedTokens = expected.Tokens();
   const std::deque<std::string>& tokens = obtained.Tokens();
@@ -457,9 +457,9 @@ TEST_CASE("StringEncodingDictionarySerialization", "[StringEncodingTest]")
 
   for (const string& line : stringEncodingInput)
   {
-    STRING_VIEW lineView(line);
+    MLPACK_STRING_VIEW lineView(line);
 
-    STRING_VIEW token = tokenizer(lineView);
+    MLPACK_STRING_VIEW token = tokenizer(lineView);
 
     while (!tokenizer.IsTokenEmpty(token))
     {
@@ -544,7 +544,7 @@ TEST_CASE("CharExtractDictionaryEncodingSerialization", "[StringEncodingTest]")
  */
 TEST_CASE("BagOfWordsEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   arma::mat output;
   BagOfWordsEncoding<SplitByAnyOf::TokenType> encoder;
@@ -619,7 +619,7 @@ TEST_CASE("BagOfWordsEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("VectorBagOfWordsEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   vector<vector<size_t>> output;
   BagOfWordsEncoding<SplitByAnyOf::TokenType> encoder(
@@ -713,7 +713,7 @@ TEST_CASE("VectorBagOfWordsEncodingIndividualCharactersTest",
  */
 TEST_CASE("RawCountSmoothIdfEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   arma::mat output;
   TfIdfEncoding<SplitByAnyOf::TokenType> encoder;
@@ -815,7 +815,7 @@ TEST_CASE("RawCountSmoothIdfEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("VectorRawCountSmoothIdfEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   vector<vector<double>> output;
   TfIdfEncoding<SplitByAnyOf::TokenType> encoder(
@@ -974,7 +974,7 @@ TEST_CASE("VectorRawCountSmoothIdfEncodingIndividualCharactersTest",
  */
 TEST_CASE("TfIdfRawCountEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   arma::mat output;
   TfIdfEncoding<SplitByAnyOf::TokenType> encoder(
@@ -1024,7 +1024,7 @@ TEST_CASE("TfIdfRawCountEncodingTest", "[StringEncodingTest]")
  */
 TEST_CASE("VectorTfIdfRawCountEncodingTest", "[StringEncodingTest]")
 {
-  using DictionaryType = StringEncodingDictionary<STRING_VIEW>;
+  using DictionaryType = StringEncodingDictionary<MLPACK_STRING_VIEW>;
 
   vector<vector<double>> output;
   TfIdfEncoding<SplitByAnyOf::TokenType>
